refactor(server): extract reference project tools into server_reference_tools module
