+++ tach.toml
@@ -51,6 +51,16 @@ depends_on = [
     { path = "mcp_workspace.file_tools" },
     { path = "mcp_workspace.git_operations" },
     { path = "mcp_workspace.reference_projects" },
+    { path = "mcp_workspace.server_reference_tools" },
+    { path = "mcp_coder_utils.log_utils" },
+]
+
+[[modules]]
+path = "mcp_workspace.server_reference_tools"
+layer = "protocol"
+depends_on = [
+    { path = "mcp_workspace.file_tools" },
+    { path = "mcp_workspace.reference_projects" },
     { path = "mcp_coder_utils.log_utils" },
 ]
 
@@ -133,6 +143,7 @@ depends_on = [
     { path = "mcp_workspace" },
     { path = "mcp_workspace.main" },
     { path = "mcp_workspace.server" },
+    { path = "mcp_workspace.server_reference_tools" },
     { path = "mcp_workspace.file_tools" },
     { path = "mcp_workspace.git_operations" },
     { path = "mcp_workspace.github_operations" },
